fix(storage): persist SQLite rowids across reload; make REPLACE INTO durable (#5891)
* fix(storage): persist SQLite rowids across reload; make REPLACE INTO durable

Fixes the corruption-class bug where rowid semantics silently changed
across a process restart (issue #5835), and the related REPLACE INTO
durability hole (issue #5871).

Root causes:
- TableSchema::rowid_alias_column / ColumnSchema::is_exact_integer_type
  were never serialized nor rebuilt on binary load, so a reloaded
  INTEGER PRIMARY KEY table forgot its rowid alias: WHERE rowid=N
  returned zero rows, DELETE ... WHERE rowid=N hit the WRONG row
  (intpkey-2.6), and RETURNING rowid yielded NULL.
- Row::row_id was not persisted (format v12) or WAL-replayed (WAL v2),
  so explicit rowids were lost and implicit rowids were renumbered by
  physical position whenever tombstones were dropped at save time; new
  inserts could then collide with reloaded rowids.
- The REPLACE conflict-delete emitted no WAL op, and REPLACE never
  counted as a modification for the CLI exit checkpoint, so the next
  process resurrected the old row next to the new one — duplicate
  INTEGER PRIMARY KEY rows (check-13.x, upsert5-3.x).

Fix:
- Rehydrate the IPK rowid alias from the persisted sql_source on binary
  load (exact-INTEGER single-column PK only; INT PRIMARY KEY excluded).
- Binary format v13: persist each live row's effective rowid (alias IPK
  value, else explicit rowid, else physical position + 1). v12 files
  stay readable.
- WAL format v3: Insert ops carry the effective rowid; recovery stamps
  it back on replay, and replayed UPDATEs preserve the target rowid.
  v2 logs stay readable.
- Table::next_rowid() tracks max(physical count, max assigned rowid) so
  allocation never collides after a reload; compaction materializes
  surviving implicit rowids before shifting positions.
- WAL-log REPLACE conflict deletes before applying them; treat REPLACE
  as a modification statement for the CLI exit checkpoint.

Tests: rowid_reload_tests.rs (binary reload + WAL crash-recovery
round-trips), two cross-process CLI tests in wal_recovery_test.rs,
format-v13 persistence tests, and Table rowid unit tests.
intpkey.test: 52 passed / 32 failed -> 84 passed / 0 failed.
check.test: 44 passed / 48 failed -> 61 passed / 31 failed (13.x clear).

Closes #5835
Closes #5871


* fix(storage): treat rowids as signed i64 so negative rowids never poison allocation

Addresses the blocking judge finding on PR #5891: an explicit negative
rowid (e.g. INSERT INTO t(rowid,x) VALUES(-1,5)) was stored as its
two's-complement u64 bit pattern (u64::MAX) and tracked as an UNSIGNED
max, so the next implicit insert computed u64::MAX + 1 — an
attempt-to-add-with-overflow panic in debug builds, and a wrap to
duplicate rowid 0 in release builds. The same poisoning applied on
reload of a negative INTEGER PRIMARY KEY (v13 persists the alias value
via `*v as u64`), where a subsequent REPLACE INTO hit the identical
overflow through its next_rowid() reservation.

Fix: adopt SQLite's signed-rowid model end-to-end while keeping
Row::row_id as the u64 bit pattern (which already round-trips
persistence and SELECT/WHERE bit-faithfully):

- Table::max_assigned_rowid is now Option<i64> — the SIGNED max over
  every effective rowid ever assigned (explicit rowids as `rid as i64`,
  implicit rows as physical position + 1, subsuming the old
  physical-count component). None = no row ever inserted.
- next_rowid_signed() = signed max + 1 (saturating at i64::MAX), or 1
  for a never-populated table — sqlite3-verified: after rowid -1 the
  next implicit rowid is 0; after -5 it is -4. next_rowid() returns the
  bit pattern.
- Batch NULL/DEFAULT rowid auto-assignment (apply_affinity) tracks the
  signed batch max, matching sqlite3 for VALUES(-1,1),(NULL,2) → -1, 0.
- The REPLACE reserved-rowid skip steps in signed space
  ((reserved as i64) + 1), so a reserved rowid of -1 steps to 0 instead
  of overflowing past u64 range.
- Rowid comparisons used for ordering now compare in signed space
  (ORDER BY rowid fast path, sort tie-breakers): rowid -1 sorts before
  0, identical to u64 order for all non-negative rowids.

Verified against sqlite3 (allocation, batch NULL assignment, ORDER BY,
UPDATE/DELETE by negative rowid, REPLACE on negative IPK) and across
save/reload: WHERE rowid=-1 finds the row after reopen and the next
implicit insert still gets 0. intpkey.test stays 84/0; check.test,
insert.test, triggerC.test unchanged vs branch baseline.

Regression tests: storage unit tests (poisoning, saturation), v13
round-trip tests (negative explicit rowid, negative IPK alias +
next_rowid sanity), and SQL-level tests for the judge repro, the batch
NULL variant, and REPLACE on a negative INTEGER PRIMARY KEY.


---------

+++ crates/vibesql-executor/src/select/executor/fast_path/helpers.rs
@@ -377,7 +377,12 @@ impl SelectExecutor<'_> {
                                 }
                             }
                             (Some(id_a), Some(id_b)) => {
-                                let raw_cmp = id_a.cmp(&id_b);
+                                // Rowids are SIGNED (issue #5835): row_id is
+                                // the two's-complement bit pattern of an i64,
+                                // so compare in signed space — rowid -1
+                                // (u64::MAX) sorts before 0. Identical to
+                                // u64 order for non-negative rowids.
+                                let raw_cmp = (id_a as i64).cmp(&(id_b as i64));
                                 match dir {
                                     OrderDirection::Asc => raw_cmp,
                                     OrderDirection::Desc => raw_cmp.reverse(),
